- Added additional APIFlowState extension functions - moved the extension functions into the APIFlowState.kt file
