+++ BaseRepo-S3Uploader/README.md
@@ -14,9 +14,9 @@ An extension module that bridges BaseRepo and S3Uploader, enabling seamless file
 
 ```gradle.kts
 // Requires both the base modules
-implementation("com.github.appoly.AppolyDroid-Toolbox:BaseRepo:1.0.24")
-implementation("com.github.appoly.AppolyDroid-Toolbox:S3Uploader:1.0.24")
-implementation("com.github.appoly.AppolyDroid-Toolbox:BaseRepo-S3Uploader:1.0.24")
+implementation("com.github.appoly.AppolyDroid-Toolbox:BaseRepo:1.0.25")
+implementation("com.github.appoly.AppolyDroid-Toolbox:S3Uploader:1.0.25")
+implementation("com.github.appoly.AppolyDroid-Toolbox:BaseRepo-S3Uploader:1.0.25")
 ```
 
 ## How it Works
