+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ composeExtensionsMinSdk = "21"
 dateHelperMinSdk = "26"
 lazyPagingMinSdk = "21"
 s3UploaderMinSdk = "21"
-toolboxVersion = "1.0.24"
+toolboxVersion = "1.0.25"
 agp = "8.12.0"
 kotlin = "2.2.0"
 ksp = "2.2.0-2.0.2"
